pyproject.toml
[tool.poetry]
name = "tastebuddy"
version = "0.1.0"
description = "TasteBuddy helps with nutrition choices and is an AI agent"
authors = ["Vasilije Markovic"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = {git = "https://github.com/topoteretes/langchain.git" , tag = "v0.0.209"}
nltk = "3.8.1"
openai = "0.27.8"
pinecone-client = "2.2.1"
python-dotenv = "1.0.0"
pyyaml = "6.0"
fastapi = "0.68.2"
uvicorn = "0.15.0"
googlemaps = "4.10.0"
jinja2 = "3.1.2"
replicate = "^0.8.1"
pexpect = "^4.8.0"
selenium = "^4.9.0"
playwright = "^1.32.1"
pytest-playwright = "^0.3.3"
boto3 = "^1.26.125"
gptcache = "^0.1.22"
redis = "^4.5.5"
gunicorn = "^20.1.0"
tiktoken = "^0.4.0"
google-search-results = "^2.4.2"
spacy = "^3.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
